fix: fix main menu button sometimes turning into a white square
This is done by adding the "windowsMenu@20x20" to ignored paths for the MainSvgPatcher.
The main problem is in the BigIconsPatcher which messes the icon up when used,
but we cannot easily disable just that one patcher for it.
We can only disable all the patchers instead.
However, this isn't as bad as it sounds, since currently none of the patchers affect the icon anyway
(it is already in grayscale and the size isn't affected when the settings change).

This is also hard to test for, since the IDE appears to have some internal cache for the icons, so old changed icons sometimes will appear not fixed on the old version or not broken on the old version.
The easiest way I found to test it was to put custom icon size to 24 and custom line height to 50.
Then I went down one line height at a time on the new version and compared that with going down one line height from 40 on the old version.
That way there is no internal icon cache, and it can be observed that after this fix the icon stays as it should, instead of turning to a white square.